chore: rename detachedvouchers to allvouchersprintout
